Extract vendored fish completion for Cobra to a temporary package
